More cleaning up. Moved pattern files to the audio sd card
It makes more sense to have a primarily read only sd card, so
the patterns directory has been moved over to the same sd card as
the audio files. The default debug.txt file was changed to be a
little more useful as well. Added MemoryFree as a dependent library.
I HAVE NO IDEA WHY. To be honest, I don't care because the broken
things appeared to become less broken with the inclusion of that
library and it's 2:00AM.