Fixed LEDs getting stuck in an undefined state
I was able to duplicate the LED undefined state by quickly unplugging
and plugging the device power back int. By setting the LEDs to 0 (dark)
during LED init, the LEDs will turn off. Note that this doesn't prevent
the LEDs from getting into the weird state, but it does appear to fix
it.